pyproject.toml
[project]
name = "giant-eagle-receipt-lookup"
version = "1.0.0"
description = "AI-powered receipt lookup for Giant Eagle on Azure Databricks"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
    "databricks-sdk>=0.20.0",
    "mlflow>=2.10.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "databricks-connect>=15.0.0",
]
pipelines = [
    "dlt",
    "pyspark>=3.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
